[Refactor] : classpath 변경으로 인한 refactor
